projects/capstone-project/deployment/dog_breed_classification/requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dataclasses==0.7
fastapi==0.55.1
future==0.18.2
google-api-core==1.19.1
google-auth==1.16.1
google-cloud-core==1.3.0
google-cloud-logging==1.15.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
h11==0.9.0
httptools==0.1.1
idna==2.9
numpy==1.18.5
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.5.1
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
starlette==0.13.2
torch==1.5.0
torchvision==0.6.0
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
